feat: add createRoomHandling.ts
