+++ src/requestsHandling/regHandling/regHandling.ts
@@ -1,38 +1,16 @@
-import { dataBase } from '../../data/data';
-import { TYPES_ENUM } from '../../interfaces/types.model';
-import { IUserCredentials, IDataReg, IReg, IRegCheck } from './reg.model';
-import { isExistedUser } from './isExistedUser';
+import WebSocket from 'ws';
+import { getRegResponse } from './getRegResponse';
+import { updateWinners } from './updateWinners';
+import { createRoomHandling } from '../roomHandling/createRoomHandling';
+import { IMessage } from '../message.model';
 
-export function regHandling(request: IReg): IRegCheck {
-  const credentials: IUserCredentials = JSON.parse(request.data);
-  const data: IDataReg = {
-    name: credentials.name,
-    index: 0,
-    error: false,
-  };
-  const existedUser = dataBase.users.find((existedUser, index) =>
-    isExistedUser(existedUser, credentials, data, index)
-  );
-
-  const response: IReg = {
-    type: TYPES_ENUM.REG,
-    data: JSON.stringify(data),
-    id: 0,
-  };
-
-  if (existedUser && existedUser.password === credentials.password) {
-    return { error: false, response: response };
-  } else {
-    return {
-      error: true,
-      response: {
-        ...response,
-        data: JSON.stringify({
-          ...data,
-          error: true,
-          errorText: 'Incorrect login or password',
-        }),
-      },
-    };
-  }
+export function regHandling(ws: WebSocket, request: IMessage) {
+  const result = getRegResponse(request);
+  ws.send(JSON.stringify(result.response));
+  !result.error
+    ? (() => {
+        ws.send(JSON.stringify(updateWinners(result.response)));
+        ws.send(JSON.stringify(createRoomHandling()));
+      })()
+    : null;
 }
